- major bug-fix for file seeking and RT grabbing - fix for video OSD status messages

+++ HomerMultimedia/include/MediaSource.h
@@ -327,7 +327,7 @@ protected:
     AVFormatContext     *mFormatContext;
     AVCodecContext      *mCodecContext;
     int                 mMediaStreamIndex;
-    int64_t             mStartPtsUSecs /* only used in MediasourceNet and MediaSourceFile */, mSourceStartPts; // for synchronized playback we calculate the position within a media stream and write the value into PTS entry of an encoded packet
+    double              mStartPtsUSecs /* only used in MediasourceNet and MediaSourceFile */, mSourceStartPts; // for synchronized playback we calculate the position within a media stream and write the value into PTS entry of an encoded packet
     int64_t             mNumberOfFrames;
     /* audio */
     AVFifoBuffer        *mRecorderSampleFifo;
